+++ Mailer/extensions/mailer-module/src/Repositories/ConfigsRepository.php
@@ -33,6 +33,9 @@ class ConfigsRepository extends Repository
         return $result;
     }
 
+    /**
+     * @deprecated Flag `autoload` will be removed in the next major release and consequently this method as well. Use `ConfigsRepository::all()` instead.
+     */
     public function loadAllAutoload(): Selection
     {
         return $this->getTable()->where('autoload', true)->order('sorting');